+++ browser/src/jsMain/kotlin/client/Actuals.kt
@@ -24,8 +24,8 @@ actual fun alert(message: String) {
     window.alert(message)
 }
 
-actual suspend fun awaitAnimationFrame(): Double = window.awaitAnimationFrame()
-actual fun getDevicePixelRatio(): Double = window.devicePixelRatio
+actual suspend fun awaitAnimationFrame(): Float = window.awaitAnimationFrame().toFloat()
+actual fun getDevicePixelRatio(): Float = window.devicePixelRatio.toFloat()
 actual fun getLocationHost(): String = window.location.host
 actual val windowAdapter: WindowAdapter = WindowAdapterImpl()
 
@@ -41,18 +41,18 @@ actual fun uint8ArrayAdapterFactory(length: Int): Uint8ArrayAdapter = Uint8Array
 actual fun audioAdapterFactory(src: String): AudioAdapter = AudioAdapterImpl(src)
 
 private val shellProgressElement = document.getElementById(SHELLS_ID) as HTMLProgressElement
-actual fun setShellsStatusBar(percent: Double) {
-    shellProgressElement.value = percent
+actual fun setShellsStatusBar(percent: Float) {
+    shellProgressElement.value = percent.toDouble()
 }
 
 private val armorProgressElement = document.getElementById(ARMOR_ID) as HTMLProgressElement
-actual fun setArmorStatusBar(percent: Double) {
-    armorProgressElement.value = percent
+actual fun setArmorStatusBar(percent: Float) {
+    armorProgressElement.value = percent.toDouble()
 }
 
 private val minesProgressElement = document.getElementById(MINES_ID) as HTMLProgressElement
-actual fun setMinesStatusBar(percent: Double) {
-    minesProgressElement.value = percent
+actual fun setMinesStatusBar(percent: Float) {
+    minesProgressElement.value = percent.toDouble()
 }
 
 private val builderModeTreeElement = document.getElementById(BUILDER_MODE_TREE_ID) as HTMLInputElement
